backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic==2.11.7
python-dotenv==1.1.1
alembic==1.16.5
pytest==8.4.2
httpx==0.28.1
PyYAML==6.0.3
